Move methods from AsyncWriteExt to AsyncWrite
