+++ examples/hello_world.rs
@@ -13,7 +13,7 @@
 
 #![warn(rust_2018_idioms)]
 
-use tokio::io::AsyncWriteExt;
+use tokio::io::AsyncWrite;
 use tokio::net::TcpStream;
 
 use std::error::Error;
