+++ tests-integration/tests/process_stdio.rs
@@ -1,7 +1,7 @@
 #![warn(rust_2018_idioms)]
 #![cfg(feature = "full")]
 
-use tokio::io::{AsyncBufReadExt, AsyncWriteExt, BufReader};
+use tokio::io::{AsyncBufReadExt, AsyncWrite, BufReader};
 use tokio::process::{Child, Command};
 use tokio_test::assert_ok;
