+++ tokio-test/tests/io.rs
@@ -1,7 +1,7 @@
 #![warn(rust_2018_idioms)]
 
 use std::io;
-use tokio::io::{AsyncRead, AsyncWriteExt};
+use tokio::io::{AsyncRead, AsyncWrite};
 use tokio_test::io::Builder;
 
 #[tokio::test]
