+++ tokio/src/io/stderr.rs
@@ -20,12 +20,12 @@ cfg_io_std! {
     ///
     /// [`stderr`]: stderr()
     /// [`AsyncWrite`]: AsyncWrite
-    /// [`write_all`]: crate::io::AsyncWriteExt::write_all()
+    /// [`write_all`]: crate::io::AsyncWrite::write_all()
     ///
     /// # Examples
     ///
     /// ```
-    /// use tokio::io::{self, AsyncWriteExt};
+    /// use tokio::io::{self, AsyncWrite};
     ///
     /// #[tokio::main]
     /// async fn main() -> io::Result<()> {
@@ -51,12 +51,12 @@ cfg_io_std! {
     /// [`write_all`] may result in interleaved output.
     ///
     /// [`AsyncWrite`]: AsyncWrite
-    /// [`write_all`]: crate::io::AsyncWriteExt::write_all()
+    /// [`write_all`]: crate::io::AsyncWrite::write_all()
     ///
     /// # Examples
     ///
     /// ```
-    /// use tokio::io::{self, AsyncWriteExt};
+    /// use tokio::io::{self, AsyncWrite};
     ///
     /// #[tokio::main]
     /// async fn main() -> io::Result<()> {
