+++ tokio/src/net/unix/stream.rs
@@ -497,7 +497,7 @@ impl AsyncWrite for UnixStream {
     }
 
     fn poll_shutdown(self: Pin<&mut Self>, _: &mut Context<'_>) -> Poll<io::Result<()>> {
-        self.shutdown(std::net::Shutdown::Write)?;
+        UnixStream::shutdown(&self, std::net::Shutdown::Write)?;
         Poll::Ready(Ok(()))
     }
 }
