+++ tokio/src/prelude.rs
@@ -17,5 +17,5 @@ pub use crate::io::{self, AsyncBufRead, AsyncRead, AsyncWrite};
 
 cfg_io_util! {
     #[doc(no_inline)]
-    pub use crate::io::{AsyncBufReadExt as _, AsyncSeekExt as _, AsyncWriteExt as _};
+    pub use crate::io::{AsyncBufReadExt as _, AsyncSeekExt as _};
 }
