+++ tokio/tests/async_send_sync.rs
@@ -319,25 +319,25 @@ async_assert_fn!(tokio::io::AsyncRead::read_i128_le(&mut BoxAsyncRead): !Unpin);
 async_assert_fn!(tokio::io::AsyncRead::read_to_end(&mut BoxAsyncRead, &mut Vec<u8>): !Unpin);
 async_assert_fn!(tokio::io::AsyncRead::read_to_string(&mut BoxAsyncRead, &mut String): !Unpin);
 async_assert_fn!(tokio::io::AsyncSeekExt::seek(&mut BoxAsyncSeek, SeekFrom): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write(&mut BoxAsyncWrite, &[u8]): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_all(&mut BoxAsyncWrite, &[u8]): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_u8(&mut BoxAsyncWrite, u8): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_i8(&mut BoxAsyncWrite, i8): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_u16(&mut BoxAsyncWrite, u16): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_i16(&mut BoxAsyncWrite, i16): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_u32(&mut BoxAsyncWrite, u32): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_i32(&mut BoxAsyncWrite, i32): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_u64(&mut BoxAsyncWrite, u64): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_i64(&mut BoxAsyncWrite, i64): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_u128(&mut BoxAsyncWrite, u128): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_i128(&mut BoxAsyncWrite, i128): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_u16_le(&mut BoxAsyncWrite, u16): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_i16_le(&mut BoxAsyncWrite, i16): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_u32_le(&mut BoxAsyncWrite, u32): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_i32_le(&mut BoxAsyncWrite, i32): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_u64_le(&mut BoxAsyncWrite, u64): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_i64_le(&mut BoxAsyncWrite, i64): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_u128_le(&mut BoxAsyncWrite, u128): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::write_i128_le(&mut BoxAsyncWrite, i128): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::flush(&mut BoxAsyncWrite): !Unpin);
-async_assert_fn!(tokio::io::AsyncWriteExt::shutdown(&mut BoxAsyncWrite): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write(&mut BoxAsyncWrite, &[u8]): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_all(&mut BoxAsyncWrite, &[u8]): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_u8(&mut BoxAsyncWrite, u8): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_i8(&mut BoxAsyncWrite, i8): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_u16(&mut BoxAsyncWrite, u16): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_i16(&mut BoxAsyncWrite, i16): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_u32(&mut BoxAsyncWrite, u32): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_i32(&mut BoxAsyncWrite, i32): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_u64(&mut BoxAsyncWrite, u64): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_i64(&mut BoxAsyncWrite, i64): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_u128(&mut BoxAsyncWrite, u128): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_i128(&mut BoxAsyncWrite, i128): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_u16_le(&mut BoxAsyncWrite, u16): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_i16_le(&mut BoxAsyncWrite, i16): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_u32_le(&mut BoxAsyncWrite, u32): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_i32_le(&mut BoxAsyncWrite, i32): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_u64_le(&mut BoxAsyncWrite, u64): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_i64_le(&mut BoxAsyncWrite, i64): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_u128_le(&mut BoxAsyncWrite, u128): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::write_i128_le(&mut BoxAsyncWrite, i128): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::flush(&mut BoxAsyncWrite): !Unpin);
+async_assert_fn!(tokio::io::AsyncWrite::shutdown(&mut BoxAsyncWrite): !Unpin);
