+++ tokio/tests/io_mem_stream.rs
@@ -1,7 +1,7 @@
 #![warn(rust_2018_idioms)]
 #![cfg(feature = "full")]
 
-use tokio::io::{duplex, AsyncRead, AsyncWriteExt};
+use tokio::io::{duplex, AsyncRead, AsyncWrite};
 
 #[tokio::test]
 async fn ping_pong() {
