+++ tokio/tests/tcp_split.rs
@@ -5,7 +5,7 @@ use std::io::Result;
 use std::io::{Read, Write};
 use std::{net, thread};
 
-use tokio::io::{AsyncRead, AsyncWriteExt};
+use tokio::io::{AsyncRead, AsyncWrite};
 use tokio::net::TcpStream;
 
 #[tokio::test]
